Moved tunnel tests to seperate package
