+++ Tribler/Test/Community/Tunnel/FullSession/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains code for TunnelCommunity tests that are starting a complete Tribler session.
+"""
